# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Smart Office monitoring API — simulated device state, power usage, and alerts
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type RoomId = typeof RoomId[keyof typeof RoomId];


export const RoomId = {
  drawing_room: 'drawing_room',
  work_room_1: 'work_room_1',
  work_room_2: 'work_room_2',
} as const;

export type DeviceType = typeof DeviceType[keyof typeof DeviceType];


export const DeviceType = {
  fan: 'fan',
  light: 'light',
} as const;

export interface Device {
  id: string;
  /** Human readable label, e.g. "Fan 1" */
  name: string;
  type: DeviceType;
  room: RoomId;
  isOn: boolean;
  /** Wattage drawn when on; 0 when off */
  powerDrawWatts: number;
  /** Rated wattage of the device when running */
  ratedWatts: number;
  lastChangedAt: string;
}

export interface DeviceUpdate {
  isOn: boolean;
}

export interface ErrorResponse {
  error: string;
}

export interface DeviceListResponse {
  devices: Device[];
  generatedAt: string;
}

export interface RoomSummary {
  room: RoomId;
  devices: Device[];
  totalWatts: number;
  activeDeviceCount: number;
  deviceCount: number;
}

export interface RoomUsage {
  room: RoomId;
  watts: number;
  activeDeviceCount: number;
  deviceCount: number;
}

export interface UsageSummary {
  totalWatts: number;
  estimatedDailyKwh: number;
  activeDeviceCount: number;
  totalDeviceCount: number;
  perRoom: RoomUsage[];
  generatedAt: string;
}

export type AlertType = typeof AlertType[keyof typeof AlertType];


export const AlertType = {
  after_hours: 'after_hours',
  prolonged_use: 'prolonged_use',
} as const;

export interface Alert {
  id: string;
  type: AlertType;
  room: RoomId;
  message: string;
  createdAt: string;
  deviceIds: string[];
}

export interface AlertListResponse {
  alerts: Alert[];
  generatedAt: string;
}