# lib/api-zod/src/generated/types/deviceUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Smart Office monitoring API — simulated device state, power usage, and alerts
 * OpenAPI spec version: 0.1.0
 */

export interface DeviceUpdate {
  isOn: boolean;
}